configure sentry for background tasks, drop elastic APM
