+++ peachjam/settings.py
@@ -247,20 +247,6 @@ REST_FRAMEWORK = {
     "PAGE_SIZE": 10,
 }
 
-# Elastic APM
-APM_SERVER_URL = os.environ.get("APM_SERVER_URL", "")
-ELASTIC_APM = {
-    "SERVICE_NAME": os.environ.get("APM_SERVICE_NAME", PEACHJAM["APP_NAME"]),
-    "SERVER_URL": APM_SERVER_URL,
-}
-if not DEBUG and APM_SERVER_URL:
-    INSTALLED_APPS = INSTALLED_APPS + ["elasticapm.contrib.django"]
-    MIDDLEWARE = [
-        "elasticapm.contrib.django.middleware.TracingMiddleware",
-        "elasticapm.contrib.django.middleware.Catch404Middleware",
-    ] + MIDDLEWARE
-
-
 # Sentry
 if not DEBUG:
     sentry_logging = LoggingIntegration(
@@ -272,7 +258,7 @@ if not DEBUG:
         environment=PEACHJAM["SENTRY_ENVIRONMENT"],
         integrations=[DjangoIntegration(), sentry_logging],
         send_default_pii=True,
-        traces_sample_rate=0.5,  # sample 50% of the requests for performance metrics
+        traces_sample_rate=0.5,  # sample 50% of requests for performance metrics
     )
 
 DEBUG_TOOLBAR_CONFIG = {"INTERCEPT_REDIRECTS": False}
